refactor: CLI driver and PyQt6 UI cleanups
- nav_mosaic: route --log-level through the program loggers, not the
  stdlib root logger (CODE-MAIN-002).
- nav_create_bundle/nav_backplane_viewer: use load_default_and_user_config
  (CODE-MAIN-005).
- nav_backplane_viewer: share _alpha_blend_layer in the on-screen path
  (CODE-UI-001).
- zoom: delegate _zoom_at_point/_zoom_in/_zoom_out to ZoomPanController
  across the three windows (CODE-UI-003/004).
- mosaic_viewer: hoist _SyncedSlider/_ZoomSync to common, public
  slider/zoom map helpers, pdslogger, annotate __all__, histogram
  tie-break fix (CODE-UI-005/006/007/009).
- annotation: clear font-load error (CODE-ANNO-2).
