+++ src/nav/ui/mosaic_viewer/common.py
@@ -6,8 +6,8 @@ reprojection / mosaic dataclasses (``RingReprojResult``, ``RingMosaicData``,
 ``DisplayData`` object ready for use in the ring or body window.
 """
 
-import logging
 import math
+from collections.abc import Callable
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import cast
@@ -16,13 +16,125 @@ import astropy.io.fits as pyfits
 import numpy as np
 import numpy.ma as ma
 from filecache import FCPath
+from PyQt6.QtWidgets import QLineEdit, QSlider
 
+from nav.config import IMAGE_LOGGER
 from nav.reproj._serialization import infer_format
 from nav.reproj.bodies import BodyMosaicData, BodyReprojResult
 from nav.reproj.ring_orbit_model import RingOrbitModel, get_orbit_model_by_name
 from nav.reproj.rings import RingMosaicData, RingReprojResult
+from nav.ui.mosaic_viewer.tiled_image_widget import slider_to_zoom, zoom_to_slider
 
-logger = logging.getLogger(__name__)
+logger = IMAGE_LOGGER
+
+
+class _SyncedSlider:
+    """Keeps a QLineEdit and QSlider in sync for a single numeric parameter."""
+
+    def __init__(
+        self,
+        line_edit: QLineEdit,
+        slider: QSlider,
+        lo: float,
+        hi: float,
+        fmt: str = '%.4f',
+        on_change: Callable[[float], None] | None = None,
+    ) -> None:
+        """Wire ``line_edit`` and ``slider`` over ``[lo, hi]`` with optional ``on_change``.
+
+        Parameters:
+            line_edit: Numeric text field.
+            slider: 0..1000 slider mapped linearly to ``[lo, hi]``.
+            lo: Lower bound of the mapped range.
+            hi: Upper bound of the mapped range.
+            fmt: ``printf`` format for the line edit.
+            on_change: Called with the new float after slider or edit updates.
+        """
+        self._le = line_edit
+        self._sl = slider
+        self._lo = lo
+        self._hi = hi
+        self._fmt = fmt
+        self._on_change = on_change
+        self._updating = False
+        self._sl.valueChanged.connect(self._slider_moved)
+        self._le.editingFinished.connect(self._edit_done)
+
+    def _to_slider(self, val: float) -> int:
+        """Map ``val`` in ``[lo, hi]`` to a slider position in ``[0, 1000]``."""
+        if self._hi <= self._lo:
+            return 0
+        pos = (val - self._lo) / (self._hi - self._lo) * 1000.0
+        return round(float(np.clip(pos, 0, 1000)))
+
+    def _from_slider(self, pos: int) -> float:
+        """Map slider position ``pos`` (0..1000) back to a float in ``[lo, hi]``."""
+        return self._lo + (self._hi - self._lo) * pos / 1000.0
+
+    def _slider_moved(self, pos: int) -> None:
+        """Mirror slider motion into the line edit and invoke ``on_change``."""
+        if self._updating:
+            return
+        val = self._from_slider(pos)
+        self._updating = True
+        self._le.setText(self._fmt % val)
+        self._updating = False
+        if self._on_change:
+            self._on_change(val)
+
+    def _edit_done(self) -> None:
+        """Parse the line edit, clamp to ``[lo, hi]``, sync the slider, call ``on_change``."""
+        if self._updating:
+            return
+        try:
+            val = float(self._le.text())
+        except ValueError:
+            return
+        val = max(self._lo, min(self._hi, val))
+        self._updating = True
+        self._sl.setValue(self._to_slider(val))
+        self._le.setText(self._fmt % val)
+        self._updating = False
+        if self._on_change:
+            self._on_change(val)
+
+    def set_range(self, lo: float, hi: float) -> None:
+        """Update ``lo``/``hi`` without changing the displayed value."""
+        self._lo = lo
+        self._hi = hi
+
+    def set_value(self, val: float) -> None:
+        """Programmatically set both widgets to ``val`` (clamped by current range)."""
+        self._updating = True
+        self._sl.setValue(self._to_slider(val))
+        self._le.setText(self._fmt % val)
+        self._updating = False
+
+    def get_value(self) -> float:
+        """Return the line-edit float if valid, otherwise infer from the slider."""
+        try:
+            return float(self._le.text())
+        except ValueError:
+            return self._from_slider(self._sl.value())
+
+
+class _ZoomSync(_SyncedSlider):
+    """A :class:`_SyncedSlider` that uses logarithmic zoom mapping.
+
+    Converts between the zoom float value and a 0-1000 slider integer using
+    :func:`~nav.ui.mosaic_viewer.tiled_image_widget.zoom_to_slider` and
+    :func:`~nav.ui.mosaic_viewer.tiled_image_widget.slider_to_zoom` so that
+    zooming feels perceptually uniform.
+    """
+
+    def _to_slider(self, val: float) -> int:
+        """Convert zoom value to slider integer position via ``zoom_to_slider``."""
+        return zoom_to_slider(val)
+
+    def _from_slider(self, pos: int) -> float:
+        """Convert slider integer position to zoom value via ``slider_to_zoom``."""
+        return slider_to_zoom(pos)
+
 
 _RAD_TO_DEG: float = 180.0 / math.pi
